+++ packages/keyring-eth-trezor/src/trezor-keyring.ts
@@ -462,12 +462,14 @@ export class TrezorKeyring implements Keyring {
   async signTypedData<
     Version extends SignTypedDataVersion.V3 | SignTypedDataVersion.V4,
     Types extends MessageTypes,
-    Options extends { version?: Version },
+    Options extends { version: Version },
   >(
     address: Hex,
     data: TypedMessage<Types>,
-    { version }: Options,
+    options?: Options,
   ): Promise<string> {
+    const { version } = options ?? { version: SignTypedDataVersion.V4 };
+
     const dataWithHashes = transformTypedData(
       data,
       version === SignTypedDataVersion.V4,